changed to pyproject.toml - added test data - renamed test files
