+++ .gitignore
@@ -1,25 +1,37 @@
-cromwell-executions/
-cromwell-workflow-logs/
+
+# built files
 __pycache__/
+build/
 .idea/
 .pyc
-.env
 .DS_Store
 *.egg-info/
 pip-wheel-metadata/
-.coverage
-.mypy_cache/
 dist/
-venv/
+
+# env & IDE
+.env
 .python-version
+.mypy_cache/
 .vscode/
-output/
+venv/
+
+# tests
+.coverage
+.nextflow/
+cromwell-executions/
+cromwell-workflow-logs/
+sample_data/
 local_data/
-docs/
-translated*
-*.log
-test.py
+
+# translation files
+output/
+translated*/
 testout/
+*.log
+involucro
+
+# misc files
+test.py
 ubu*
-janis_assistant/tests/data
-sample_data
+temp.txt
