+++ pyproject.toml
@@ -5,7 +5,7 @@ requires        = ["setuptools>=67.8.0", "wheel>=0.40.0"]
 
 [project]
 name            = "janis-pipelines.assistant"
-version         = "v0.13.0-beta1"
+version         = "v0.13.0"
 description     = "Easier way to run workflows, configurable across environments"
 readme          = "README.md"
 license         = { file = "LICENSE" }
@@ -38,7 +38,8 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Bio-Informatics",
 ]
 dependencies = [
-    "janis-pipelines.core @ git+https://github.com/GraceAHall/janis-core@release-v0.13.0-beta1",
+    # "janis-pipelines.core @ git+https://github.com/GraceAHall/janis-core@release-v0.13.0",
+    "../janis-core",
     "janis-pipelines.templates >= 0.11.0",
     "requests",
     "path",
@@ -65,19 +66,19 @@ ci = [
     "twine",
 ]
 
-[project.scripts]
-janis = "janis_assistant.cli:process_args"
-
-[project.entry-points."janis.extension"]
-assistant = "janis_assistant"
-
 [project.urls]
 repository      = "https://github.com/PMCC-BioinformaticsCore/janis-assistant"
 documentation   = "https://janis.readthedocs.io/en/latest/"
 
 [tool.setuptools.packages.find]
-where = ["./janis_assistant"]
+where = ["./"]
+include = ["janis_assistant*"]
+namespaces = false
 
-[tool.setuptools]
-include-package-data = true
+[project.scripts]
+janis = "janis_assistant.cli:process_args"
 
+
+# NOTE
+# The following line was not copied across from old setup.py (unknown purpose)
+# entry_points={"janis.extension": ["assistant=janis_assistant"]}
