--- temp.txt
@@ -1,65 +1,0 @@
-
-
-photo 
-
-title
-
-abstract 
- - the background of the talk. 
- - Could be to give a brief description of their company (what, why, how) and their role
-
-Melbourne Bioinformatics 
-- Training (workshops, MC-SCIBIF)
-- Connection: linking you to bioinformatian with experience in your field
-- Advice on research design, data analysis 
-- Bioinformatics collaboration / services for research projects
-- Developing & administrating software platforms - Galaxy, Janis, other biocommons projects
-
-
-Software Engineering within Bioinformatics - Pathways, Current State, Future Directions
-
-I am a Bioinformatician & Software Engineer at Melbourne Bioinformatics. 
-In this Q&A talk, I will discuss the pathway which lead me to my current position. 
-I will also share some insight into the type of work performed by software engineers within bioinformatics, 
-as well as the current state of Research Software Engineers (RSEs) in this field. 
-To finish, I will briefly share my thoughts on the future direciton of RSE work within bioinformatics, and will hazard a guess as to how AI will affect this field.
-
-Melbourne Bioinformatics provides a range of functions at the University of Melbourne. 
-Aside from providing training workshops, contributing to the MC-SCIBIF course, and our involvement in the Galaxy project, we provide bioinformatics services to research groups. This involves consultation on research design and the data analysis required for a given project, as well as collaborating on projects if needed. Melbourne bioinformatics also functions to connect researchers to an external bioinformatian with experience in their field. 
-
-
-
-How does my story apply to other people? 
-- Do what you like to do 
-- This will mean you're dedicated & motivated
-- Lots of evidence to support that becoming an 'expert' is about time committed, rather than innate ability. 
-
-AI?
-- Research is actually a good place to be. AI tooling will make our work more productive.
-- As a software engineer, I would be more in the firing line, but you still need to know what you're trying to accomplish. 
-- At the end of the day, software engineers ship features & products, not code. 
-- User doesn't care about the code. Its your job as a RSE to use any tools available to generate code which accomplishes your goal. 
-- Might compare it to c vs python - single line of python might be dozens if not hundreds of c. 
-- People use python because it accomplishes their goal faster than c. 
-- My manager recently said something that stuck :
-- "the goal of research is to put yourself out of a job"
-- yet research still exists.
-- For the forseeable future, will still need people to guide research efforts. 
-
-
-
-
-
-talk
-- University background
-- Pathway (how university + personal study lead to first job)
-- Job history (tutoring / lecturing, previous role @ melbourne bioinformatics, current role @ melbourne bioinformatics)
-- My thoughts on professional life as an academic bioinformatician - networking, type of jobs available etc
-- Masters vs PhD 
-- Industry vs Academia 
-- AI
-
-
-
-
-
